backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
django-image-uploader-widget==1.0.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
mccabe==0.7.0
pillow==11.0.0
pycodestyle==2.12.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
